45 seconds -> 18 seconds
All of the speedup comes from not adding j to i's neighbours if j < i
while building the neighbour sets. This avoids having to check it
multiple times while clique finding, and both the neighbour and clique
loops become a bit more than twice as fast.

I have also removed the enumeration and subsequent index lookups. The
bit sets are inherently hashable and sortable, so can be used directly
in place of the indices. This has no noticable effect on speed but it
does make the code slightly easier to understand. :)

Finally I changed alphaBit to use str.lower() instead of manual lower
casing. This is about 3% faster, but building the bitsets takes less
than a second already.